% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefClassPopulation.R
\name{assignPoD}
\alias{assignPoD}
\title{Assign probability of disease (PoD)}
\arguments{
\item{x}{numeric vector - vector of estimated PoD values}
}
\value{
Subject level probability of disease for the population
}
\description{
Function assigns subject-level probability of disease based on PoD curve and subject level titer.
}
\details{
The input into the function is either calculated using \code{PoD} function or if the PoD curve is unknown the same arbitrary PoD can be assigned to the whole population.
}
