% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{efficacySet}
\alias{efficacySet}
\title{Estimated PoDBAY efficacies}
\format{
vector
\describe{
  \item{numeric vector}{PoDBAY efficacies}
}
}
\usage{
efficacySet
}
\description{
A dataset containing estimated set of PoDBAY efficacies. (Set of efficacies is a vector obtained by number of replications specified by repeatCount. These replications are performed for calculation of a confidence interval. For more details, see the supplementary material of the article).
}
\keyword{datasets}
