% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estimatedParameters}
\alias{estimatedParameters}
\title{Estimated PoD curve parameters}
\format{
data frame
\describe{
  \item{pmax}{pmax: maximum PoD}
  \item{et50}{et50: titer value corresponding to the pmax/2}
  \item{slope}{slope: slope of the PoD curve}
}
}
\usage{
estimatedParameters
}
\description{
A dataset containing estimated set of PoD curve parameters. (Set of PoD curve parameters is a vector obtained by number of replications specified by repeatCount. These replications are performed for calculation of a confidence interval. For more details, see the supplementary material of the article).
}
\keyword{datasets}
