% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefClassPopulation.R
\name{getNondiseasedCount}
\alias{getNondiseasedCount}
\title{Non-diseased count}
\value{
numeric: number of the non-diseased subjects in the \code{Population-class} object
}
\description{
Function calculates the number of non-diseased subjects (disease status = FALSE) in the \code{Population-class} object.
}
\details{
Input into the function, "diseaseStatus", is taken from the \code{Population-class} object attribute. Information about disease status is written into the \code{Population-class} object by the \code{ClinicalTrial()} function.
}
