% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefClassPopulation.R
\name{getTiters}
\alias{getTiters}
\title{Subject level titers}
\value{
Subject level titers
}
\description{
Returns subject level titers. If titers are not yet generated, the function generates them based on \code{Population-class} object attributes: N, mean, stdDev.
}
\details{
Inputs into the function (N, mean, stdDev) are taken from the \code{Population-class} object attributes.
}
