% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefClassPopulation.R
\name{getUnknown}
\alias{getUnknown}
\title{Generate unknown}
\arguments{
\item{n}{numeric: number of subjects in the population}
}
\value{
unknown part of the titers
}
\description{
Function generates unknown part of the titers which is eventually added to the original titers in \code{popX} and to the original titer distribution in \code{popFun}.
}
\details{
Input into the function: UDFunction is taken from the \code{Population-class} object. UDFunction is used for generating the unknown part of the titer distribution.
}
