% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsSupplementary.R
\name{incorrectInput}
\alias{incorrectInput}
\title{Error message}
\usage{
incorrectInput(name)
}
\arguments{
\item{name}{name of the input value}
}
\value{
error message: "the input value for "name" is incorrect"
}
\description{
Error meassage: the input value for "name" is incorrent
}
