% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsSupplementary.R
\name{incorrectPopulationInput}
\alias{incorrectPopulationInput}
\title{Population class error message}
\usage{
incorrectPopulationInput(name)
}
\arguments{
\item{name}{name of the input value}
}
\value{
error message: "The input value for "name" is incorrect. Input needs to be a population class object."
}
\description{
Error meassage: the input value for "name" is incorrect.
}
