% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPODBAYEfficacy.R
\name{JitterMean}
\alias{JitterMean}
\title{Population mean jittering}
\usage{
JitterMean(blindPopulation)
}
\arguments{
\item{blindPopulation}{\code{Population-class} object with N, mean, stdDev attributes}
}
\value{
Jittered mean, numeric value
}
\description{
Function jitters the mean of the population.

Jittering is adding noise to the mean. The jittered mean is sampled from the distribution with the population mean and population standard deviation divided by the number of subjects in the population. The input population is provided in the form of population class objects (see the \code{Population-class} function for more details).

\deqn{Mean_{jitter} \sim N(mean, \frac{sd}{N} )}{ MeanJitter ~ N(mean, sd/N)}
}
\examples{
## Data preparation
data(vaccinated)

## Example 1
vaccinated$mean
JitterMean(vaccinated)

}
