% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PODBAY.R
\docType{package}
\name{PoDBAY}
\alias{PoDBAY}
\title{PoDBAY}
\description{
PoDBAY package accompanies the article 'A method to estimate probability
of disease and vaccine efficacy 
from clinical trial immunogenicity data'. 
It helps to setup the workflow for vaccine efficacy estimation and clinical 
trial simulation using the PoDBAY method.
}
\details{
It has two main applications:

\itemize{
 \item Estimation of vaccine efficacy using subject level immunogenicity data
 \item Simulation of clinical trial 
}
}
\author{
Pavel Fiser, Tomas Bartonek, Julie Dudasova
}
