% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPoDParameters.R
\name{PoDCI}
\alias{PoDCI}
\title{PoD curve confidence ribbon}
\usage{
PoDCI(data, ci = 0.95)
}
\arguments{
\item{data}{numeric vector for which we the confidence intervals should be calculated}

\item{ci}{numeric: required confidence level}
}
\value{
lower bound of CI
median value
upper bound of CI
}
\description{
Supplementary function for \code{PoDCurvePlot} function. Function calculates the confidence ribbon around the PoD curve.
}
\examples{
## Data preparation
data <- 0:100

## Example 1
PoDCI(data,
      ci = 0.95)

}
