% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PoDParams}
\alias{PoDParams}
\title{PoD curve parameters}
\format{
data frame
\describe{
  \item{pmax}{pmax: maximum PoD}
  \item{et50}{et50: titer value corresponding to the pmax/2}
  \item{slope}{slope: slope of the PoD curve}
}
}
\usage{
PoDParams
}
\description{
A dataset containing PoD curve parameters
}
\keyword{datasets}
