% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsPoDParameters.R
\name{PoDParamsCICoverage}
\alias{PoDParamsCICoverage}
\title{Confidence intervals of PoD curve parameters at three confidence levels}
\usage{
PoDParamsCICoverage(estimatedParameters, ci = 0.95)
}
\arguments{
\item{estimatedParameters}{output of \code{PoDParamEstimation} function}

\item{ci}{numeric: value from (0, 1) interval, confidence level of interest}
}
\value{
CI of all PoD curve parameters
}
\description{
Function calculates confidence intervals (80\%, 90\% and user-defined) of the PoD curve parameters (pmax, et50, slope).
}
