% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefClassPopulation.R
\name{popFun}
\alias{popFun}
\title{Population function}
\value{
Titer distribution function
}
\description{
Function describing the titer distribution of the population: mean, standard deviation and an additional unknown factor affecting the shape of the distribution (e.g. mixture of two normals or other shapes defined by user).
}
\details{
Inputs into the function (mean, stdDev, Unknowndistribution) and getUnknown method are taken from the \code{Population-class} object.
}
