% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefClassPopulation.R
\docType{class}
\name{Population-class}
\alias{Population-class}
\alias{population}
\title{Population class}
\description{
Population reference class which provides summary and subject level information about the population
}
\section{Fields}{

\describe{
\item{\code{N}}{numeric: number of subjects in the population}

\item{\code{mean}}{numeric: mean value of titers}

\item{\code{stdDev}}{numeric: standard deviation of titers}

\item{\code{unknownDistribution}}{logical: TRUE if titer distribution is not normally /log-normally distributed; titer disrtibution function needs to be defined by user}

\item{\code{UDFunction}}{function: user-defined titer distribution}

\item{\code{titers}}{numeric: subject level titers, generated with \code{getTiters} method}

\item{\code{PoDs}}{numeric: subject level probability of disease (PoD), generated with \code{assginPoD} method}

\item{\code{diseaseStatus}}{logical: subject level disease status (TRUE if diseased), generated with \code{ClinicilaTrial} function}
}}


