% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefClassPopulation.R
\name{popX}
\alias{popX}
\title{Add noise to population titers}
\value{
subject level titers
}
\description{
Function adds noise to population titers accounting for an unknown factor affecting the titer distibution.
}
\details{
Inputs into the function: N, unknownDistribution and getUnknown() method are taken from the \code{Population-class} object.
}
