% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vaccinated}
\alias{vaccinated}
\title{Dataset containing the information for vaccinated subjects}
\format{
Population class object:
\describe{
  \item{N}{number of subjects}
  \item{mean}{mean of titers}
  \item{stdDev}{standard deviation of titers}
  \item{titers}{subject level titers}
}
}
\usage{
vaccinated
}
\description{
A dataset containing the N, mean, stdDev, titers of vaccinated subjects. The dataset is provided in the form of population class object (see the \code{Population-class} function for more details).
}
\keyword{datasets}
